Fix FrameBufferVideoRenderer so that it can be attached to any producer
Ignore-this: 8f6002de26491a1d761c4453d5dce560

Before this change it was possible to change the producer on the fly, but
only to another producer of the same type. A VideoCamera is an active
producer, whereas a RemoteVideoStream is a passive one. Now the renderer
will destroy the underlying video port and recreate it automagically, in
order to accomodate any producer type.

darcs-hash:20140924111242-56b4d-b37bef3f6f4c7c6ee3e672c88754843dc938eaaf