Reject incoming re-INVITE if we couldn't reply to it
Ignore-this: a160cb766e08c47ad4055dd340521313

Only do it if it wasn't a stream removal, that is, if the remote party
tried to change some session / stream parameters for which we couldn't
build an answer. This is a best effort approach.

darcs-hash:20141117175048-56b4d-99cf842168eb2a471707c2c87781e544c03833a0